requirements.txt
python-dotenv==1.0.1
slack-bolt==1.19.1
slack-sdk==3.27.1
Flask==3.0.3
requests==2.32.3
httpx==0.27.2
tenacity==8.4.2
pytest==8.3.2

